+++ packages/core/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.2](https://github.com/vtex/faststore/compare/v4.1.1...v4.1.2) (2026-06-01)
+
+### Bug Fixes
+
+- **core:** fix locale validation for disabled localization and pvt routes ([#3358](https://github.com/vtex/faststore/issues/3358)) ([b742658](https://github.com/vtex/faststore/commit/b742658140d8b3cd5a42767a3eab1ef64ce0e63e))
+
 ## 4.1.1 (2026-05-15)
 
 ### Bug Fixes